refactor(asset-server-plugin): Improve organization of source files
