+++ packages/asset-server-plugin/index.ts
@@ -1,6 +1,8 @@
 export * from './src/plugin';
-export * from './src/s3-asset-storage-strategy';
-export * from './src/sharp-asset-preview-strategy';
-export * from './src/image-transform-strategy';
-export * from './src/preset-only-strategy';
+export * from './src/asset-server';
+export * from './src/config/s3-asset-storage-strategy';
+export * from './src/config/sharp-asset-preview-strategy';
+export * from './src/config/image-transform-strategy';
+export * from './src/config/preset-only-strategy';
+export * from './src/config/hashed-asset-naming-strategy';
 export * from './src/types';
