+++ packages/asset-server-plugin/src/config/hashed-asset-naming-strategy.ts
