+++ packages/asset-server-plugin/src/config/local-asset-storage-strategy.ts
