+++ packages/asset-server-plugin/src/config/preset-only-strategy.ts
@@ -1,9 +1,10 @@
+import { ImageTransformFormat } from '../types';
+
 import {
     GetImageTransformParametersArgs,
     ImageTransformParameters,
     ImageTransformStrategy,
 } from './image-transform-strategy';
-import { ImageTransformFormat } from './types';
 
 /**
  * @description
